Improve container detection
Rely on 'container' environment variable and '/.dockerenv'-file to determine
if build should pivot to container. The 'container' variable is used by
sensible container runtimes, such as podman and lxc, while docker creates
'/.dockerenv' file to container root. Use combination of the to
determine if container should be entered

As a bonus, this adds support to native builds: by setting 'container'
variable f.ex. to 'skip', entering container is skipped.

Signed-off-by: Markku Ahvenjärvi <markkux@ssrc.tii.ae>