Podman support, quality fixes.
Implement good support for using Podman as an
alternative to Docker.
-Avoid UID/GID problems with '--userns=keep-id'
-Add 'CAP_SYS_MODULE' capability to Podman container.
 In some cases the build system(s) use 'binfmt' kernel
 module, and SELinux blocks containers from inserting
 kernel modules by default.

Refactor passing the SSH keys to container. With Podman,
use Podman secrets, and with Docker, the SSH keys are
copied to a temporary directory for passthrough. Rationale
behind this is to avoid SELinux labeling issues, which is
described in detail in 'utils.sh'.

Add 'utils.sh' for helper functions, and for generating
runtime arguments.

Add Bash helper script which is added to the container image
during build. This script handles finding the SSH keys passed
through to the container, and adds them to the SSH agent.

Signed-off-by: Joonas Onatsu <joonasx@ssrc.tii.ae>